requirements.txt
Json
request
speech_recognition
Pyttsx3
Wikipedia
Ecapture
time